🔧 Refactor: Introduce centralized error handling and update route setup for improved clarity and maintainability
