+++ services/catalog/src/simcore_service_catalog/api/rest/errors.py
@@ -1,0 +1,6 @@
+from fastapi import FastAPI
+from servicelib.fastapi.http_error import set_app_default_http_error_handlers
+
+
+def setup_rest_exception_handlers(app: FastAPI) -> None:
+    set_app_default_http_error_handlers(app)
