+++ services/catalog/src/simcore_service_catalog/api/rest/routes.py
@@ -14,67 +14,67 @@ from . import (
     _services_specifications,
 )
 
-v0_router = APIRouter()
 
-# health
-health_router = _health.router
-v0_router.include_router(
-    _health.router,
-    tags=["diagnostics"],
-)
+def setup_rest_routes(app: FastAPI, vtag: str):
+    v0_router = APIRouter()
 
-# meta
-v0_router.include_router(
-    _meta.router,
-    tags=["meta"],
-    prefix="/meta",
-)
+    # health
+    health_router = _health.router
+    v0_router.include_router(
+        _health.router,
+        tags=["diagnostics"],
+    )
 
-# services
-_SERVICE_PREFIX = "/services"
-_SERVICE_TAGS: list[str | Enum] = [
-    "services",
-]
-v0_router.include_router(
-    _services_resources.router,
-    tags=_SERVICE_TAGS,
-    prefix=_SERVICE_PREFIX,
-)
-v0_router.include_router(
-    _services_labels.router,
-    tags=_SERVICE_TAGS,
-    prefix=_SERVICE_PREFIX,
-)
-v0_router.include_router(
-    _services_extras.router,
-    tags=_SERVICE_TAGS,
-    prefix=_SERVICE_PREFIX,
-)
-v0_router.include_router(
-    _services_specifications.router,
-    tags=_SERVICE_TAGS,
-    prefix=_SERVICE_PREFIX,
-)
-v0_router.include_router(
-    _services_ports.router,
-    tags=_SERVICE_TAGS,
-    prefix=_SERVICE_PREFIX,
-)
-v0_router.include_router(
-    _services_access_rights.router,
-    tags=_SERVICE_TAGS,
-    prefix=_SERVICE_PREFIX,
-)
+    # meta
+    v0_router.include_router(
+        _meta.router,
+        tags=["meta"],
+        prefix="/meta",
+    )
 
-# NOTE: that this router must come after resources/specifications/ports/access_rights
-v0_router.include_router(
-    _services.router,
-    tags=_SERVICE_TAGS,
-    prefix=_SERVICE_PREFIX,
-)
+    # services
+    _SERVICE_PREFIX = "/services"
+    _SERVICE_TAGS: list[str | Enum] = [
+        "services",
+    ]
+    v0_router.include_router(
+        _services_resources.router,
+        tags=_SERVICE_TAGS,
+        prefix=_SERVICE_PREFIX,
+    )
+    v0_router.include_router(
+        _services_labels.router,
+        tags=_SERVICE_TAGS,
+        prefix=_SERVICE_PREFIX,
+    )
+    v0_router.include_router(
+        _services_extras.router,
+        tags=_SERVICE_TAGS,
+        prefix=_SERVICE_PREFIX,
+    )
+    v0_router.include_router(
+        _services_specifications.router,
+        tags=_SERVICE_TAGS,
+        prefix=_SERVICE_PREFIX,
+    )
+    v0_router.include_router(
+        _services_ports.router,
+        tags=_SERVICE_TAGS,
+        prefix=_SERVICE_PREFIX,
+    )
+    v0_router.include_router(
+        _services_access_rights.router,
+        tags=_SERVICE_TAGS,
+        prefix=_SERVICE_PREFIX,
+    )
 
+    # NOTE: that this router must come after resources/specifications/ports/access_rights
+    v0_router.include_router(
+        _services.router,
+        tags=_SERVICE_TAGS,
+        prefix=_SERVICE_PREFIX,
+    )
 
-def setup_rest_api_routes(app: FastAPI, vtag: str):
     # healthcheck at / and at /v0/
     app.include_router(health_router)
     # api under /v*
