+++ services/catalog/src/simcore_service_catalog/api/rpc/routes.py
@@ -11,7 +11,7 @@ from . import _services
 _logger = logging.getLogger(__name__)
 
 
-async def setup_rpc_api_routes(app: FastAPI) -> AsyncIterator[State]:
+async def setup_rpc_routes(app: FastAPI) -> AsyncIterator[State]:
     rpc_server = get_rabbitmq_rpc_server(app)
     await rpc_server.register_router(_services.router, CATALOG_RPC_NAMESPACE, app)
